+++ Assets/Scripts/Buildings/BuildingsImplementation/MeatShop/MeatShopLogic.cs
@@ -133,8 +133,8 @@ public class MeatShopLogic : BuildingLogic, IWorkplace
         return baseAmount * IndexMultiplier;
     }
 
-    public override int GetBuildCost() => data.GetBuildCost(IndexMultiplier);
-    public override int GetUpgradeCost() => data.GetUpgradeCost(Level, IndexMultiplier);
+    public override long GetBuildCost() => data.GetBuildCost(IndexMultiplier);
+    public override long GetUpgradeCost() => data.GetUpgradeCost(Level, IndexMultiplier);
     public override string GetCycleLabel() => "Cycle";
     public override string GetAmountLabel() => "Meat";
     public override string GetWorkAnimation() => null;
